revm_bytecode::eof

Module verification

Verification logic for the EOF bytecode. Module that contains the verification logic for the EOF bytecode.

Structs

AccessTracker

Tracker status of verification of code sections and subcontainers. Used in validating EOF container.

Enums

CodeType

Types of code sections in EOF container

EofError

EOF Error

EofValidationError

EOF Validation Error

Functions

validate_eof

Fully validates an Eof container.

validate_eof_code

Validates that:

validate_eof_codes

Validates an Eof structure, without recursing into containers.

validate_eof_inner

Fully validates an Eof container. If first_code_type is None it will be auto deduced in verification process.

validate_raw_eof

Decodes raw into an Eof container and validates it.

validate_raw_eof_inner

Decodes raw into an Eof container and validates it.