revm_bytecode
Module legacy

Legacy bytecode.

Structs

JumpTable

A table of valid jump destinations. Cheap to clone and memory efficient, one bit per opcode.

LegacyAnalyzedBytecode

Legacy analyzed bytecode represents the original bytecode format used in Ethereum.

LegacyRawBytecode

Used only as intermediate representation for legacy bytecode. Please check LegacyAnalyzedBytecode for the main structure that is used in Revm.

Functions

analyze_legacy

Analyze the bytecode to find the jumpdests. Used to create a jump table that is needed for crate::LegacyAnalyzedBytecode. This function contains a hot loop and should be optimized as much as possible.