revm_bytecode

Module eof

EOF bytecode.

Contains body, header and raw bytes.

Also contains verification logic and pretty printer.

Re-exports

pub use verification::*;

Modules

printer

Pritty printer for the EOF bytecode. Enabled by std feature. Module that contains the printer for the EOF bytecode.

verification

Verification logic for the EOF bytecode. Module that contains the verification logic for the EOF bytecode.

Structs

CodeInfo

Types section that contains stack information for matching code section

Eof

EVM Object Format (EOF) container

EofBody

EOF container body

EofHeader

EOF header structure that contains section sizes and metadata

Enums

EofDecodeError

EOF decode errors

Constants

CODE_SECTION_SIZE

EOF header code section size length.

CONTAINER_SECTION_SIZE

EOF header container section size length.

EOF_MAGIC

EOF Magic in u16 form

EOF_MAGIC_HASH

Hash of EF00 bytes that is used for EXTCODEHASH when called from legacy bytecode

KIND_CODE

EOF header code kind, marking code section.

KIND_CODE_INFO

EOF header code info kind, marking code info section.

KIND_CONTAINER

EOF header container kind, marking container section.

KIND_DATA

EOF header data kind, marking data section.

KIND_TERMINAL

EOF header terminal kind, marking end of header.

Statics

EOF_MAGIC_BYTES

EOF magic number in array form