revm_bytecode::opcode

Struct OpCode

pub struct OpCode(/* private fields */);

An EVM opcode

This is always a valid opcode, as declared in the opcode module or the OPCODE_INFO constant.

Implementations

impl OpCode

pub const fn new(opcode: u8) -> Option<Self>

Instantiates a new opcode from a u8.

Returns None if the opcode is not valid.

pub const fn is_jumpdest(&self) -> bool

Returns true if the opcode is a jump destination.

pub const fn is_jumpdest_by_op(opcode: u8) -> bool

Takes a u8 and returns true if it is a jump destination.

pub const fn is_jump(self) -> bool

Returns true if the opcode is a legacy jump instruction.

pub const fn is_jump_by_op(opcode: u8) -> bool

Takes a u8 and returns true if it is a jump instruction.

pub const fn is_push(self) -> bool

Returns true if the opcode is a PUSH instruction.

pub fn is_push_by_op(opcode: u8) -> bool

Takes a u8 and returns true if it is a push instruction.

pub unsafe fn new_unchecked(opcode: u8) -> Self

Instantiates a new opcode from a u8 without checking if it is valid.

Safety

All code using Opcode values assume that they are valid opcodes, so providing an invalid opcode may cause undefined behavior.

pub const fn as_str(self) -> &'static str

Returns the opcode as a string. This is the inverse of parse.

pub const fn name_by_op(opcode: u8) -> &'static str

Returns the opcode name.

pub const fn inputs(&self) -> u8

Returns the number of input stack elements.

pub const fn outputs(&self) -> u8

Returns the number of output stack elements.

pub const fn io_diff(&self) -> i16

Calculates the difference between the number of input and output stack elements.

pub const fn info_by_op(opcode: u8) -> Option<OpCodeInfo>

Returns the opcode information for the given opcode. Check OpCodeInfo for more information.

pub const fn as_usize(&self) -> usize

Returns the opcode as a usize.

pub const fn info(&self) -> OpCodeInfo

Returns the opcode information.

pub const fn input_output(&self) -> (u8, u8)

Returns the number of both input and output stack elements.

Can be slightly faster that calling inputs and outputs separately.

pub const fn get(self) -> u8

Returns the opcode as a u8.

pub const fn modifies_memory(&self) -> bool

Returns true if the opcode modifies memory.

https://bluealloy.github.io/revm/crates/interpreter/memory.html#opcodes

https://github.com/crytic/evm-opcodes

impl OpCode

pub const STOP: Self

The 0x00 (“STOP”) opcode.

pub const ADD: Self

The 0x01 (“ADD”) opcode.

pub const MUL: Self

The 0x02 (“MUL”) opcode.

pub const SUB: Self

The 0x03 (“SUB”) opcode.

pub const DIV: Self

The 0x04 (“DIV”) opcode.

pub const SDIV: Self

The 0x05 (“SDIV”) opcode.

pub const MOD: Self

The 0x06 (“MOD”) opcode.

pub const SMOD: Self

The 0x07 (“SMOD”) opcode.

pub const ADDMOD: Self

The 0x08 (“ADDMOD”) opcode.

pub const MULMOD: Self

The 0x09 (“MULMOD”) opcode.

pub const EXP: Self

The 0x0A (“EXP”) opcode.

pub const SIGNEXTEND: Self

The 0x0B (“SIGNEXTEND”) opcode.

pub const LT: Self

The 0x10 (“LT”) opcode.

pub const GT: Self

The 0x11 (“GT”) opcode.

pub const SLT: Self

The 0x12 (“SLT”) opcode.

pub const SGT: Self

The 0x13 (“SGT”) opcode.

pub const EQ: Self

The 0x14 (“EQ”) opcode.

pub const ISZERO: Self

The 0x15 (“ISZERO”) opcode.

pub const AND: Self

The 0x16 (“AND”) opcode.

pub const OR: Self

The 0x17 (“OR”) opcode.

pub const XOR: Self

The 0x18 (“XOR”) opcode.

pub const NOT: Self

The 0x19 (“NOT”) opcode.

pub const BYTE: Self

The 0x1A (“BYTE”) opcode.

pub const SHL: Self

The 0x1B (“SHL”) opcode.

pub const SHR: Self

The 0x1C (“SHR”) opcode.

pub const SAR: Self

The 0x1D (“SAR”) opcode.

pub const KECCAK256: Self

The 0x20 (“KECCAK256”) opcode.

pub const ADDRESS: Self

The 0x30 (“ADDRESS”) opcode.

pub const BALANCE: Self

The 0x31 (“BALANCE”) opcode.

pub const ORIGIN: Self

The 0x32 (“ORIGIN”) opcode.

pub const CALLER: Self

The 0x33 (“CALLER”) opcode.

pub const CALLVALUE: Self

The 0x34 (“CALLVALUE”) opcode.

pub const CALLDATALOAD: Self

The 0x35 (“CALLDATALOAD”) opcode.

pub const CALLDATASIZE: Self

The 0x36 (“CALLDATASIZE”) opcode.

pub const CALLDATACOPY: Self

The 0x37 (“CALLDATACOPY”) opcode.

pub const CODESIZE: Self

The 0x38 (“CODESIZE”) opcode.

pub const CODECOPY: Self

The 0x39 (“CODECOPY”) opcode.

pub const GASPRICE: Self

The 0x3A (“GASPRICE”) opcode.

pub const EXTCODESIZE: Self

The 0x3B (“EXTCODESIZE”) opcode.

pub const EXTCODECOPY: Self

The 0x3C (“EXTCODECOPY”) opcode.

pub const RETURNDATASIZE: Self

The 0x3D (“RETURNDATASIZE”) opcode.

pub const RETURNDATACOPY: Self

The 0x3E (“RETURNDATACOPY”) opcode.

pub const EXTCODEHASH: Self

The 0x3F (“EXTCODEHASH”) opcode.

pub const BLOCKHASH: Self

The 0x40 (“BLOCKHASH”) opcode.

pub const COINBASE: Self

The 0x41 (“COINBASE”) opcode.

pub const TIMESTAMP: Self

The 0x42 (“TIMESTAMP”) opcode.

pub const NUMBER: Self

The 0x43 (“NUMBER”) opcode.

pub const DIFFICULTY: Self

The 0x44 (“DIFFICULTY”) opcode.

pub const GASLIMIT: Self

The 0x45 (“GASLIMIT”) opcode.

pub const CHAINID: Self

The 0x46 (“CHAINID”) opcode.

pub const SELFBALANCE: Self

The 0x47 (“SELFBALANCE”) opcode.

pub const BASEFEE: Self

The 0x48 (“BASEFEE”) opcode.

pub const BLOBHASH: Self

The 0x49 (“BLOBHASH”) opcode.

pub const BLOBBASEFEE: Self

The 0x4A (“BLOBBASEFEE”) opcode.

pub const POP: Self

The 0x50 (“POP”) opcode.

pub const MLOAD: Self

The 0x51 (“MLOAD”) opcode.

pub const MSTORE: Self

The 0x52 (“MSTORE”) opcode.

pub const MSTORE8: Self

The 0x53 (“MSTORE8”) opcode.

pub const SLOAD: Self

The 0x54 (“SLOAD”) opcode.

pub const SSTORE: Self

The 0x55 (“SSTORE”) opcode.

pub const JUMP: Self

The 0x56 (“JUMP”) opcode.

pub const JUMPI: Self

The 0x57 (“JUMPI”) opcode.

pub const PC: Self

The 0x58 (“PC”) opcode.

pub const MSIZE: Self

The 0x59 (“MSIZE”) opcode.

pub const GAS: Self

The 0x5A (“GAS”) opcode.

pub const JUMPDEST: Self

The 0x5B (“JUMPDEST”) opcode.

pub const TLOAD: Self

The 0x5C (“TLOAD”) opcode.

pub const TSTORE: Self

The 0x5D (“TSTORE”) opcode.

pub const MCOPY: Self

The 0x5E (“MCOPY”) opcode.

pub const PUSH0: Self

The 0x5F (“PUSH0”) opcode.

pub const PUSH1: Self

The 0x60 (“PUSH1”) opcode.

pub const PUSH2: Self

The 0x61 (“PUSH2”) opcode.

pub const PUSH3: Self

The 0x62 (“PUSH3”) opcode.

pub const PUSH4: Self

The 0x63 (“PUSH4”) opcode.

pub const PUSH5: Self

The 0x64 (“PUSH5”) opcode.

pub const PUSH6: Self

The 0x65 (“PUSH6”) opcode.

pub const PUSH7: Self

The 0x66 (“PUSH7”) opcode.

pub const PUSH8: Self

The 0x67 (“PUSH8”) opcode.

pub const PUSH9: Self

The 0x68 (“PUSH9”) opcode.

pub const PUSH10: Self

The 0x69 (“PUSH10”) opcode.

pub const PUSH11: Self

The 0x6A (“PUSH11”) opcode.

pub const PUSH12: Self

The 0x6B (“PUSH12”) opcode.

pub const PUSH13: Self

The 0x6C (“PUSH13”) opcode.

pub const PUSH14: Self

The 0x6D (“PUSH14”) opcode.

pub const PUSH15: Self

The 0x6E (“PUSH15”) opcode.

pub const PUSH16: Self

The 0x6F (“PUSH16”) opcode.

pub const PUSH17: Self

The 0x70 (“PUSH17”) opcode.

pub const PUSH18: Self

The 0x71 (“PUSH18”) opcode.

pub const PUSH19: Self

The 0x72 (“PUSH19”) opcode.

pub const PUSH20: Self

The 0x73 (“PUSH20”) opcode.

pub const PUSH21: Self

The 0x74 (“PUSH21”) opcode.

pub const PUSH22: Self

The 0x75 (“PUSH22”) opcode.

pub const PUSH23: Self

The 0x76 (“PUSH23”) opcode.

pub const PUSH24: Self

The 0x77 (“PUSH24”) opcode.

pub const PUSH25: Self

The 0x78 (“PUSH25”) opcode.

pub const PUSH26: Self

The 0x79 (“PUSH26”) opcode.

pub const PUSH27: Self

The 0x7A (“PUSH27”) opcode.

pub const PUSH28: Self

The 0x7B (“PUSH28”) opcode.

pub const PUSH29: Self

The 0x7C (“PUSH29”) opcode.

pub const PUSH30: Self

The 0x7D (“PUSH30”) opcode.

pub const PUSH31: Self

The 0x7E (“PUSH31”) opcode.

pub const PUSH32: Self

The 0x7F (“PUSH32”) opcode.

pub const DUP1: Self

The 0x80 (“DUP1”) opcode.

pub const DUP2: Self

The 0x81 (“DUP2”) opcode.

pub const DUP3: Self

The 0x82 (“DUP3”) opcode.

pub const DUP4: Self

The 0x83 (“DUP4”) opcode.

pub const DUP5: Self

The 0x84 (“DUP5”) opcode.

pub const DUP6: Self

The 0x85 (“DUP6”) opcode.

pub const DUP7: Self

The 0x86 (“DUP7”) opcode.

pub const DUP8: Self

The 0x87 (“DUP8”) opcode.

pub const DUP9: Self

The 0x88 (“DUP9”) opcode.

pub const DUP10: Self

The 0x89 (“DUP10”) opcode.

pub const DUP11: Self

The 0x8A (“DUP11”) opcode.

pub const DUP12: Self

The 0x8B (“DUP12”) opcode.

pub const DUP13: Self

The 0x8C (“DUP13”) opcode.

pub const DUP14: Self

The 0x8D (“DUP14”) opcode.

pub const DUP15: Self

The 0x8E (“DUP15”) opcode.

pub const DUP16: Self

The 0x8F (“DUP16”) opcode.

pub const SWAP1: Self

The 0x90 (“SWAP1”) opcode.

pub const SWAP2: Self

The 0x91 (“SWAP2”) opcode.

pub const SWAP3: Self

The 0x92 (“SWAP3”) opcode.

pub const SWAP4: Self

The 0x93 (“SWAP4”) opcode.

pub const SWAP5: Self

The 0x94 (“SWAP5”) opcode.

pub const SWAP6: Self

The 0x95 (“SWAP6”) opcode.

pub const SWAP7: Self

The 0x96 (“SWAP7”) opcode.

pub const SWAP8: Self

The 0x97 (“SWAP8”) opcode.

pub const SWAP9: Self

The 0x98 (“SWAP9”) opcode.

pub const SWAP10: Self

The 0x99 (“SWAP10”) opcode.

pub const SWAP11: Self

The 0x9A (“SWAP11”) opcode.

pub const SWAP12: Self

The 0x9B (“SWAP12”) opcode.

pub const SWAP13: Self

The 0x9C (“SWAP13”) opcode.

pub const SWAP14: Self

The 0x9D (“SWAP14”) opcode.

pub const SWAP15: Self

The 0x9E (“SWAP15”) opcode.

pub const SWAP16: Self

The 0x9F (“SWAP16”) opcode.

pub const LOG0: Self

The 0xA0 (“LOG0”) opcode.

pub const LOG1: Self

The 0xA1 (“LOG1”) opcode.

pub const LOG2: Self

The 0xA2 (“LOG2”) opcode.

pub const LOG3: Self

The 0xA3 (“LOG3”) opcode.

pub const LOG4: Self

The 0xA4 (“LOG4”) opcode.

pub const DATALOAD: Self

The 0xD0 (“DATALOAD”) opcode.

pub const DATALOADN: Self

The 0xD1 (“DATALOADN”) opcode.

pub const DATASIZE: Self

The 0xD2 (“DATASIZE”) opcode.

pub const DATACOPY: Self

The 0xD3 (“DATACOPY”) opcode.

pub const RJUMP: Self

The 0xE0 (“RJUMP”) opcode.

pub const RJUMPI: Self

The 0xE1 (“RJUMPI”) opcode.

pub const RJUMPV: Self

The 0xE2 (“RJUMPV”) opcode.

pub const CALLF: Self

The 0xE3 (“CALLF”) opcode.

pub const RETF: Self

The 0xE4 (“RETF”) opcode.

pub const JUMPF: Self

The 0xE5 (“JUMPF”) opcode.

pub const DUPN: Self

The 0xE6 (“DUPN”) opcode.

pub const SWAPN: Self

The 0xE7 (“SWAPN”) opcode.

pub const EXCHANGE: Self

The 0xE8 (“EXCHANGE”) opcode.

pub const EOFCREATE: Self

The 0xEC (“EOFCREATE”) opcode.

pub const TXCREATE: Self

The 0xED (“TXCREATE”) opcode.

pub const RETURNCONTRACT: Self

The 0xEE (“RETURNCONTRACT”) opcode.

pub const CREATE: Self

The 0xF0 (“CREATE”) opcode.

pub const CALL: Self

The 0xF1 (“CALL”) opcode.

pub const CALLCODE: Self

The 0xF2 (“CALLCODE”) opcode.

pub const RETURN: Self

The 0xF3 (“RETURN”) opcode.

pub const DELEGATECALL: Self

The 0xF4 (“DELEGATECALL”) opcode.

pub const CREATE2: Self

The 0xF5 (“CREATE2”) opcode.

pub const RETURNDATALOAD: Self

The 0xF7 (“RETURNDATALOAD”) opcode.

pub const EXTCALL: Self

The 0xF8 (“EXTCALL”) opcode.

pub const EXTDELEGATECALL: Self

The 0xF9 (“EXTDELEGATECALL”) opcode.

pub const STATICCALL: Self

The 0xFA (“STATICCALL”) opcode.

pub const EXTSTATICCALL: Self

The 0xFB (“EXTSTATICCALL”) opcode.

pub const REVERT: Self

The 0xFD (“REVERT”) opcode.

pub const INVALID: Self

The 0xFE (“INVALID”) opcode.

pub const SELFDESTRUCT: Self

The 0xFF (“SELFDESTRUCT”) opcode.